+++ src/api-legacy/v1/borehole/profile/producer.py
@@ -4,7 +4,6 @@ from bms.v1.handlers import Producer
 from bms.v1.borehole.stratigraphy import (
     AddBedrock,
     CreateStratigraphy,
-    PatchStartigraphy,
     DeleteStratigraphy,
     CloneStratigraphy
 )