+++ src/api-legacy/v1/borehole/project/__init__.py
@@ -1,3 +1,0 @@
-
-# Actions
-from bms.v1.borehole.project.list import ListProject