+++ src/client/src/commons/detail/detailsState.js
@@ -1,10 +1,8 @@
 const initialState = {
   error: null,
   isFetching: false,
-  isFetchingStratigraphies: false,
   tab: 0,
   borehole: null,
-  stratigraphies: [],
 };
 
 const detail_borehole = (state = initialState, action) => {
@@ -20,7 +18,6 @@ const detail_borehole = (state = initialState, action) => {
         ...state,
         error: null,
         borehole: null,
-        stratigraphies: [],
         isFetching: true,
       };
     }
@@ -36,20 +33,6 @@ const detail_borehole = (state = initialState, action) => {
         ...state,
         isFetching: false,
         error: action.error,
-      };
-    }
-    case "GET_BOREHOLE_STRATIGRAPHIES": {
-      return {
-        ...state,
-        stratigraphies: [],
-        isFetchingStratigraphies: true,
-      };
-    }
-    case "GET_BOREHOLE_STRATIGRAPHIES_OK": {
-      return {
-        ...state,
-        isFetchingStratigraphies: false,
-        stratigraphies: action.stratigraphies,
       };
     }
     default: