+++ src/client/src/commons/search/editor/searchEditorState.js
@@ -140,8 +140,10 @@ const initialState = {
     top_bedrock_tvd_to: "",
     qt_depth: null,
     qt_total_depth_tvd: null,
-    qt_top_bedrock: null,
-    qt_top_bedrock_tvd: null,
+    qt_top_bedrock_from: "",
+    qt_top_bedrock_to: "",
+    qt_top_bedrock_tvd_from: "",
+    qt_top_bedrock_tvd_to: "",
 
     reference_elevation_type: null,
     qt_reference_elevation: null,